feat: :zap: module refactor and documentation fix
