+++ sword/src/core/application/builder.rs
@@ -25,24 +25,24 @@ use crate::{
 };
 
 /// Builder for constructing a Sword application with various configuration options.
-/// 
+///
 /// `ApplicationBuilder` provides a fluent interface for configuring a Sword application
 /// before building the final `Application` instance. It allows you to register
 /// controllers, add middleware layers, configure shared state, and set up dependency injection.
-/// 
+///
 /// # Example
-/// 
+///
 /// ```rust,ignore
 /// use sword::prelude::*;
-/// 
+///
 /// #[derive(Default)]
 /// struct AppState {
 ///     counter: std::sync::atomic::AtomicU64,
 /// }
-/// 
+///
 /// #[controller]
 /// struct HomeController;
-/// 
+///
 /// let app = Application::builder()?
 ///     .with_state(AppState::default())?
 ///     .with_controller::<HomeController>()
@@ -61,19 +61,19 @@ pub struct ApplicationBuilder {
 
 impl ApplicationBuilder {
     /// Creates a new application builder with default configuration.
-    /// 
+    ///
     /// This method initializes a new builder with:
     /// - Empty router
     /// - Fresh state container
     /// - Configuration loaded from `config/config.toml`
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns `Ok(ApplicationBuilder)` if initialization succeeds, or
     /// `Err(ApplicationError)` if configuration loading fails.
-    /// 
+    ///
     /// # Errors
-    /// 
+    ///
     /// This function will return an error if:
     /// - The configuration file cannot be found or read
     /// - The TOML syntax is invalid
@@ -94,35 +94,35 @@ impl ApplicationBuilder {
     }
 
     /// Registers a controller in the application.
-    /// 
+    ///
     /// This method adds a controller and its routes to the application's router.
     /// Controllers must implement the `RouterProvider` trait, which is typically
     /// done using the `#[controller]` and `#[routes]` macros.
-    /// 
+    ///
     /// # Type Parameters
-    /// 
+    ///
     /// * `R` - A type implementing `RouterProvider` that defines the controller's routes
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns `Self` for method chaining.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust,ignore
     /// use sword::prelude::*;
-    /// 
+    ///
     /// #[controller]
     /// struct HomeController;
-    /// 
+    ///
     /// #[routes]
     /// impl HomeController {
     ///     #[get("/")]
-    ///     async fn index() -> HttpResult<&'static str> {
-    ///         Ok("Hello, World!")
+    ///     async fn index() -> HttpResult<HttpResponse> {
+    ///         Ok(HttpResponse::Ok().message("Welcome to the Home Page"))
     ///     }
     /// }
-    /// 
+    ///
     /// let app = Application::builder()?
     ///     .with_controller::<HomeController>()
     ///     .build();
@@ -139,30 +139,30 @@ impl ApplicationBuilder {
     }
 
     /// Registers a middleware layer in the application.
-    /// 
+    ///
     /// This method allows you to add Tower-based middleware or other layers
     /// that implement the `Layer` trait. Layers are applied to all routes
     /// in the application and can modify requests and responses.
-    /// 
+    ///
     /// # Type Parameters
-    /// 
+    ///
     /// * `L` - A type implementing the `Layer` trait
-    /// 
+    ///
     /// # Arguments
-    /// 
+    ///
     /// * `layer` - The middleware layer to add to the application
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns `Self` for method chaining.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust,ignore
     /// use sword::prelude::*;
     /// use tower_http::cors::CorsLayer;
     /// use tower_http::trace::TraceLayer;
-    /// 
+    ///
     /// let app = Application::builder()?
     ///     .with_layer(CorsLayer::permissive())
     ///     .with_layer(TraceLayer::new_for_http())
@@ -186,48 +186,48 @@ impl ApplicationBuilder {
     }
 
     /// Registers shared state in the application.
-    /// 
+    ///
     /// This method adds shared state that can be accessed by controllers and middleware
     /// throughout the application. The state is automatically wrapped in an `Arc` for
     /// safe sharing across multiple threads.
-    /// 
+    ///
     /// State can be retrieved in handlers using `Context::get_state::<T>()`.
-    /// 
+    ///
     /// # Type Parameters
-    /// 
+    ///
     /// * `S` - The type of state to store (must implement `Sync + Send + 'static`)
-    /// 
+    ///
     /// # Arguments
-    /// 
+    ///
     /// * `state` - The state instance to store in the application
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns `Ok(Self)` for method chaining, or `Err(StateError)` if the state
     /// type is already registered.
-    /// 
+    ///
     /// # Errors
-    /// 
+    ///
     /// This function will return an error if the same state type has already
     /// been registered in the application.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust,ignore
     /// use sword::prelude::*;
     /// use std::sync::atomic::AtomicU64;
-    /// 
+    ///
     /// #[derive(Default)]
     /// struct AppState {
     ///     counter: AtomicU64,
     ///     name: String,
     /// }
-    /// 
+    ///
     /// let app_state = AppState {
     ///     counter: AtomicU64::new(0),
     ///     name: "My App".to_string(),
     /// };
-    /// 
+    ///
     /// let app = Application::builder()?
     ///     .with_state(app_state)?
     ///     .build();
@@ -245,56 +245,56 @@ impl ApplicationBuilder {
     }
 
     /// Registers a Shaku dependency injection module in the application.
-    /// 
+    ///
     /// This method integrates Shaku modules for dependency injection, allowing you
     /// to register services and dependencies that can be resolved later using
     /// `Context::di::<ModuleType, InterfaceType>()`.
-    /// 
+    ///
     /// Available only when the `shaku-di` feature is enabled.
-    /// 
+    ///
     /// # Type Parameters
-    /// 
+    ///
     /// * `M` - The Shaku module type (must implement `Sync + Send + 'static`)
-    /// 
+    ///
     /// # Arguments
-    /// 
+    ///
     /// * `module` - The Shaku module instance containing registered services
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns `Ok(Self)` for method chaining, or `Err(StateError)` if the module
     /// type is already registered.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust,ignore
     /// use sword::prelude::*;
     /// use shaku::{module, Component, Interface};
     /// use std::sync::Arc;
-    /// 
+    ///
     /// trait DatabaseService: Interface {
     ///     fn get_connection(&self) -> String;
     /// }
-    /// 
+    ///
     /// #[derive(Component)]
     /// #[shaku(interface = DatabaseService)]
     /// struct PostgresService;
-    /// 
+    ///
     /// impl DatabaseService for PostgresService {
     ///     fn get_connection(&self) -> String {
     ///         "postgresql://localhost:5432/mydb".to_string()
     ///     }
     /// }
-    /// 
+    ///
     /// module! {
     ///     AppModule {
     ///         components = [PostgresService],
     ///         providers = []
     ///     }
     /// }
-    /// 
+    ///
     /// let module = AppModule::builder().build();
-    /// 
+    ///
     /// let app = Application::builder()?
     ///     .with_shaku_di_module(module)?
     ///     .build();
@@ -308,31 +308,31 @@ impl ApplicationBuilder {
     }
 
     /// Builds the final application instance.
-    /// 
+    ///
     /// This method finalizes the application configuration and creates the
     /// `Application` instance. It applies all configured middleware layers,
     /// sets up request body limits, and prepares the application for running.
-    /// 
+    ///
     /// # Built-in Middleware
-    /// 
+    ///
     /// The following middleware is automatically applied:
     /// - Content-Type validation middleware
     /// - Request body size limiting middleware
     /// - Cookie management layer (if `cookies` feature is enabled)
-    /// 
+    ///
     /// # Returns
-    /// 
+    ///
     /// Returns the configured `Application` instance ready to run.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust,ignore
     /// use sword::prelude::*;
-    /// 
+    ///
     /// let app = Application::builder()?
     ///     .with_controller::<HomeController>()
     ///     .build();
-    /// 
+    ///
     /// // Application is now ready to run
     /// app.run().await?;
     /// ```
